Live alerts updated to offline will show how long the stream was live for (estimated for when the alert was triggered, catchup may cause this to be wrong) Remove storing cache/callbacks in bot class, only fetching them temporarily when needed. This was a shitty fix to the race conditions, and wasn't actually the issue Fix error referencing wrong variable in webserver notification Clean up is streamer already live checking with function abstraction
All events are not pushed through an asyncio queue, to prevent race conditions with rare occurences of multiple channels being created.
Added a queue worker that will call the appropropriate functions and dispatch events.
Functions are called manually instead of with dispatch to ensure no issues with race conditions (One of those just in case things)